\name{PictureContent-class}
\docType{class}
\alias{PictureContent-class}
\title{Class \code{"PictureContent"}}
\description{
  A virtual class that is used for verifying validity of
  \code{"Picture"} objects.
}
\author{
  Simon Potter
}
\seealso{
  \code{\linkS4class{PicturePath}}, \code{\linkS4class{PictureRect}},
  \code{\linkS4class{PictureGroup}}, etc.
}
