\name{PictureSummary-class}
\docType{class}
\alias{PictureSummary-class}
\title{Class "PictureSummary" }
\description{Summary information about a picture
regarding the scales of its bounding box. }
\section{Slots}{
  \describe{
    \item{\code{xscale}:}{Object of class \code{"numeric"} 
range of x-values in picture.}
    \item{\code{yscale}:}{Object of class \code{"numeric"} 
range of y-values in picture.}
  }
}
\author{ Simon Potter }
\seealso{
\code{\link{Picture-class}}
}
\keyword{classes}
