\name{readPicture}
\alias{readPicture}
\title{ Import a Cairo SVG image }
\description{
  This function reads in a Cairo SVG image file (produced by
  \code{convertPicture} from the grConvert package) and creates a
  \code{"Picture"} object.
}
\usage{
readPicture(file, warn = TRUE)
}
\arguments{
  \item{file}{ The filename of a Cairo SVG image. }
  \item{warn}{
    Logical. If \code{TRUE}, this will warn if \code{file} was not
    created by grConvert. This is not strictly necessary if the image is
    known to be generated by Cairo.
  }
}
\value{
  An object of class \code{"Picture"}.
}
\author{ Simon Potter }
